pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "brewt"
version = "0.1.0"
description = "Python module for generating password lists from potential options"
authors = [{ name = "Gregory Boyce", email = "gregory.boyce@gmail.com" }]
requires-python = ">=3.8"
dependencies = ["python-gnupg"]

[project.urls]
Homepage = "https://github.com/gdfuego/brewt"

[project.scripts]
brewt = "brewt:main"
brewt-gpg = "brewt_gpg:main"

[tool.setuptools]
py-modules = ["brewt", "brewt_gpg"]
